Remove booking done
